# Conflicts:
#	.github/workflows/maven-test.yml
#	docker/lakesoul-docker-compose-env/docker-compose.yml
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/DBManager.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/DataCommitInfoDao.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/NamespaceDao.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/PartitionInfoDao.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/TableInfoDao.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/TableNameIdDao.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/TablePathIdDao.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/entity/DataCommitInfo.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/entity/Namespace.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/entity/PartitionInfo.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/entity/TableInfo.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/entity/TableNameId.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/entity/TablePathId.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/rbac/AuthZAfter.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/rbac/AuthZBefore.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/rbac/fetcher/AuthZFetcher.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/rbac/fetcher/AuthZIdentFetcher.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/rbac/fetcher/AuthZNullFetcher.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/rbac/fetcher/AuthZParamFetcher.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/rbac/fetcher/TableReadFetcher.java
#	lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/rbac/fetcher/TableWriteFetcher.java
#	lakesoul-common/src/test/java/com/dmetasoul/lakesoul/meta/rbac/AuthZDomainTestDataCommitTest.java
#	lakesoul-flink/pom.xml
#	lakesoul-flink/src/main/java/org/apache/flink/lakesoul/types/BinarySourceRecord.java
#	lakesoul-flink/src/test/java/com/dmetasoul/lakesoul/meta/LakeSoulRBACTest.java
#	lakesoul-spark/pom.xml
#	lakesoul-spark/src/test/scala/com/dmetasoul/lakesoul/meta/RBACOperatinSuite.scala
#	script/meta_rbac_add_domain.sql
#	script/meta_rbac_add_user.sql
#	script/meta_rbac_init.sql

+++ lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/DataCommitInfoDao.java
@@ -19,21 +19,26 @@ package com.dmetasoul.lakesoul.meta.dao;
 
 import com.dmetasoul.lakesoul.meta.DBConnector;
 import com.dmetasoul.lakesoul.meta.DBUtil;
-import com.dmetasoul.lakesoul.meta.entity.DataCommitInfo;
+import com.dmetasoul.lakesoul.meta.entity.*;
+import com.dmetasoul.lakesoul.meta.jnr.NativeMetadataJavaClient;
+import com.dmetasoul.lakesoul.meta.jnr.NativeUtils;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
 
 public class DataCommitInfoDao {
 
     public void insert(DataCommitInfo dataCommitInfo) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.insert(
+                    NativeUtils.CodedDaoType.InsertDataCommitInfo,
+                    JniWrapper.newBuilder().addDataCommitInfo(dataCommitInfo).build());
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         try {
@@ -51,6 +56,12 @@ public class DataCommitInfoDao {
     }
 
     public void deleteByPrimaryKey(String tableId, String partitionDesc, UUID commitId) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeleteOneDataCommitInfoByTableIdAndPartitionDescAndCommitId,
+                    Arrays.asList(tableId, partitionDesc, commitId.toString()));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql = "delete from data_commit_info where table_id = ? and partition_desc = ? and commit_id = ? ";
@@ -68,7 +79,15 @@ public class DataCommitInfoDao {
         }
     }
 
-    public void deleteByTableIdPartitionDescCommitList(String tableId, String partitionDesc, List<UUID> commitIdList) {
+    public void deleteByTableIdPartitionDescCommitList(String tableId, String partitionDesc, List<Uuid> commitIdList) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeleteDataCommitInfoByTableIdAndPartitionDescAndCommitIdList,
+                    Arrays.asList(tableId, partitionDesc,
+                            commitIdList.stream().map(DBUtil::protoUuidToJniString).collect(Collectors.joining(""))
+                    ));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         if (commitIdList.size() < 1) {
@@ -83,8 +102,8 @@ public class DataCommitInfoDao {
             pstmt.setString(1, tableId);
             pstmt.setString(2, partitionDesc);
             int index = 3;
-            for (UUID uuid : commitIdList) {
-                pstmt.setString(index++, uuid.toString());
+            for (Uuid uuid : commitIdList) {
+                pstmt.setString(index++, DBUtil.toJavaUUID(uuid).toString());
             }
             pstmt.execute();
         } catch (SQLException e) {
@@ -95,6 +114,12 @@ public class DataCommitInfoDao {
     }
 
     public void deleteByTableIdAndPartitionDesc(String tableId, String partitionDesc) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeleteDataCommitInfoByTableIdAndPartitionDesc,
+                    Arrays.asList(tableId, partitionDesc));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql = "delete from data_commit_info where table_id = ? and partition_desc = ?";
@@ -112,6 +137,12 @@ public class DataCommitInfoDao {
     }
 
     public void deleteByTableId(String tableId) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeleteDataCommitInfoByTableId,
+                    Collections.singletonList(tableId));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql = "delete from data_commit_info where table_id = ?";
@@ -127,7 +158,15 @@ public class DataCommitInfoDao {
         }
     }
 
-    public DataCommitInfo selectByPrimaryKey(String tableId, String partitionDesc, UUID commitId) {
+    public DataCommitInfo selectByPrimaryKey(String tableId, String partitionDesc, String commitId) {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.SelectOneDataCommitInfoByTableIdAndPartitionDescAndCommitId,
+                    Arrays.asList(tableId, partitionDesc, commitId));
+            if (jniWrapper == null) return null;
+            List<DataCommitInfo> dataCommitInfoList = jniWrapper.getDataCommitInfoList();
+            return dataCommitInfoList.isEmpty() ? null : dataCommitInfoList.get(0);
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -139,11 +178,10 @@ public class DataCommitInfoDao {
             pstmt = conn.prepareStatement(sql);
             pstmt.setString(1, tableId);
             pstmt.setString(2, partitionDesc);
-            pstmt.setString(3, commitId.toString());
+            pstmt.setString(3, commitId);
             rs = pstmt.executeQuery();
             while (rs.next()) {
-                dataCommitInfo = new DataCommitInfo();
-                createDataCommitInfoFromRs(rs, dataCommitInfo);
+                dataCommitInfo = dataCommitInfoFromResultSet(rs);
             }
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -154,6 +192,14 @@ public class DataCommitInfoDao {
     }
 
     public DataCommitInfo selectByTableId(String tableId) {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.SelectOneDataCommitInfoByTableId,
+                    Collections.singletonList(tableId));
+            if (jniWrapper == null) return null;
+            List<DataCommitInfo> dataCommitInfoList = jniWrapper.getDataCommitInfoList();
+            return dataCommitInfoList.isEmpty() ? null : dataCommitInfoList.get(0);
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -166,8 +212,7 @@ public class DataCommitInfoDao {
             pstmt = conn.prepareStatement(sql);
             rs = pstmt.executeQuery();
             while (rs.next()) {
-                dataCommitInfo = new DataCommitInfo();
-                createDataCommitInfoFromRs(rs, dataCommitInfo);
+                dataCommitInfo = dataCommitInfoFromResultSet(rs);
             }
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -178,7 +223,19 @@ public class DataCommitInfoDao {
     }
 
     public List<DataCommitInfo> selectByTableIdPartitionDescCommitList(String tableId, String partitionDesc,
-                                                                       List<UUID> commitIdList) {
+                                                                       List<Uuid> commitIdList) {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            if (commitIdList.isEmpty()) {
+                return Collections.emptyList();
+            }
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListDataCommitInfoByTableIdAndPartitionDescAndCommitList,
+                    Arrays.asList(tableId, partitionDesc,
+                            commitIdList.stream().map(DBUtil::protoUuidToJniString).collect(Collectors.joining(""))
+                    ));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getDataCommitInfoList();
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -186,7 +243,7 @@ public class DataCommitInfoDao {
         if (commitIdList.size() < 1) {
             return commitInfoList;
         }
-        String uuidListOrderString = commitIdList.stream().map(UUID::toString).collect(Collectors.joining(","));
+        String uuidListOrderString = commitIdList.stream().map(uuid -> DBUtil.toJavaUUID(uuid).toString()).collect(Collectors.joining(","));
         String sql = String.format("select * from data_commit_info where table_id = ? and partition_desc = ? and " +
                 "commit_id in (%s) order by position(commit_id::text in ?) ", String.join(",", Collections.nCopies(commitIdList.size(), "?")));
 
@@ -196,15 +253,14 @@ public class DataCommitInfoDao {
             pstmt.setString(1, tableId);
             pstmt.setString(2, partitionDesc);
             int index = 3;
-            for (UUID uuid : commitIdList) {
-                pstmt.setString(index++, uuid.toString());
+            for (Uuid uuid : commitIdList) {
+                pstmt.setString(index++, DBUtil.toJavaUUID(uuid).toString());
             }
             pstmt.setString(index, uuidListOrderString);
 
             rs = pstmt.executeQuery();
             while (rs.next()) {
-                DataCommitInfo dataCommitInfo = new DataCommitInfo();
-                createDataCommitInfoFromRs(rs, dataCommitInfo);
+                DataCommitInfo dataCommitInfo = dataCommitInfoFromResultSet(rs);
                 commitInfoList.add(dataCommitInfo);
             }
         } catch (SQLException e) {
@@ -215,18 +271,29 @@ public class DataCommitInfoDao {
         return commitInfoList;
     }
 
-    private void createDataCommitInfoFromRs(ResultSet rs, DataCommitInfo dataCommitInfo) throws SQLException {
-        dataCommitInfo.setTableId(rs.getString("table_id"));
-        dataCommitInfo.setPartitionDesc(rs.getString("partition_desc"));
-        dataCommitInfo.setCommitId(UUID.fromString(rs.getString("commit_id")));
-        dataCommitInfo.setFileOps(DBUtil.changeStringToDataFileOpList(rs.getString("file_ops")));
-        dataCommitInfo.setCommitOp(rs.getString("commit_op"));
-        dataCommitInfo.setTimestamp(rs.getLong("timestamp"));
-        dataCommitInfo.setCommitted(rs.getBoolean("committed"));
-        dataCommitInfo.setDomain(rs.getString("domain"));
+    public static DataCommitInfo dataCommitInfoFromResultSet(ResultSet rs) throws SQLException {
+        UUID uuid = UUID.fromString(rs.getString("commit_id"));
+        Uuid commitId = DBUtil.toProtoUuid(uuid);
+        return DataCommitInfo.newBuilder()
+                .setTableId(rs.getString("table_id"))
+                .setPartitionDesc(rs.getString("partition_desc"))
+                .setCommitId(commitId)
+                .addAllFileOps(DBUtil.changeStringToDataFileOpList(rs.getString("file_ops")))
+                .setCommitOp(CommitOp.valueOf(rs.getString("commit_op")))
+                .setTimestamp(rs.getLong("timestamp"))
+                .setCommitted(rs.getBoolean("committed"))
+                .setDomain(rs.getString("domain"))
+                .build();
     }
 
     public boolean batchInsert(List<DataCommitInfo> listData) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            if (listData.isEmpty()) return true;
+            Integer count = NativeMetadataJavaClient.insert(
+                    NativeUtils.CodedDaoType.TransactionInsertDataCommitInfo,
+                    JniWrapper.newBuilder().addAllDataCommitInfo(listData).build());
+            return count > 0;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         boolean result = true;
@@ -259,11 +326,11 @@ public class DataCommitInfoDao {
     private void dataCommitInsert(PreparedStatement pstmt, DataCommitInfo dataCommitInfo) throws SQLException {
         pstmt.setString(1, dataCommitInfo.getTableId());
         pstmt.setString(2, dataCommitInfo.getPartitionDesc());
-        pstmt.setString(3, dataCommitInfo.getCommitId().toString());
-        pstmt.setString(4, DBUtil.changeDataFileOpListToString(dataCommitInfo.getFileOps()));
-        pstmt.setString(5, dataCommitInfo.getCommitOp());
+        pstmt.setString(3, DBUtil.toJavaUUID(dataCommitInfo.getCommitId()).toString());
+        pstmt.setString(4, DBUtil.changeDataFileOpListToString(dataCommitInfo.getFileOpsList()));
+        pstmt.setString(5, dataCommitInfo.getCommitOp().toString());
         pstmt.setLong(6, dataCommitInfo.getTimestamp());
-        pstmt.setBoolean(7, dataCommitInfo.isCommitted());
+        pstmt.setBoolean(7, dataCommitInfo.getCommitted());
         pstmt.setString(8, dataCommitInfo.getDomain());
         pstmt.execute();
     }
