+++ lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/NamespaceDao.java
@@ -18,11 +18,12 @@
  */
 package com.dmetasoul.lakesoul.meta.dao;
 
-import com.alibaba.fastjson.JSONObject;
 import com.dmetasoul.lakesoul.meta.DBConfig;
 import com.dmetasoul.lakesoul.meta.DBConnector;
-import com.dmetasoul.lakesoul.meta.DBUtil;
+import com.dmetasoul.lakesoul.meta.entity.JniWrapper;
 import com.dmetasoul.lakesoul.meta.entity.Namespace;
+import com.dmetasoul.lakesoul.meta.jnr.NativeMetadataJavaClient;
+import com.dmetasoul.lakesoul.meta.jnr.NativeUtils;
 import com.dmetasoul.lakesoul.meta.rbac.AuthZContext;
 import com.dmetasoul.lakesoul.meta.rbac.AuthZEnforcer;
 import dev.failsafe.internal.util.Lists;
@@ -32,10 +33,19 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class NamespaceDao {
     public void insert(Namespace namespace) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.insert(
+                    NativeUtils.CodedDaoType.InsertNamespace,
+                    JniWrapper.newBuilder().addNamespace(namespace).build());
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         try {
@@ -43,7 +53,7 @@ public class NamespaceDao {
             pstmt = conn.prepareStatement("insert into namespace(namespace, properties, comment, domain) " +
                     "values (?, ?, ?, ?)");
             pstmt.setString(1, namespace.getNamespace());
-            pstmt.setString(2, DBUtil.jsonToString(namespace.getProperties()));
+            pstmt.setString(2, namespace.getProperties());
             pstmt.setString(3, namespace.getComment());
             pstmt.setString(4, namespace.getDomain());
             pstmt.execute();
@@ -55,6 +65,14 @@ public class NamespaceDao {
     }
 
     public Namespace findByNamespace(String name) {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.SelectNamespaceByNamespace,
+                    Collections.singletonList(name));
+            if (jniWrapper == null) return null;
+            List<Namespace> namespaceList = jniWrapper.getNamespaceList();
+            return namespaceList.isEmpty() ? null : namespaceList.get(0);
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -65,11 +83,7 @@ public class NamespaceDao {
             pstmt = conn.prepareStatement(sql);
             rs = pstmt.executeQuery();
             while (rs.next()) {
-                namespace = new Namespace();
-                namespace.setNamespace(rs.getString("namespace"));
-                namespace.setProperties(DBUtil.stringToJSON(rs.getString("properties")));
-                namespace.setComment(rs.getString("comment"));
-                namespace.setDomain(rs.getString("domain"));
+                namespace = namespaceFromResultSet(rs);
             }
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -80,6 +94,12 @@ public class NamespaceDao {
     }
 
     public void deleteByNamespace(String namespace) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeleteNamespaceByNamespace,
+                    Collections.singletonList(namespace));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql = String.format("delete from namespace where namespace = '%s' ", namespace);
@@ -95,6 +115,14 @@ public class NamespaceDao {
     }
 
     public List<String> listNamespaces() {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListNamespaces,
+                    Collections.emptyList());
+            if (jniWrapper == null) return null;
+            List<Namespace> namespaceList = jniWrapper.getNamespaceList();
+            return namespaceList.stream().map(Namespace::getNamespace).collect(Collectors.toList());
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -116,13 +144,18 @@ public class NamespaceDao {
         return list;
     }
 
-    public int updatePropertiesByNamespace(String namespace, JSONObject properties) {
+    public int updatePropertiesByNamespace(String namespace, String properties) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            return NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.UpdateNamespacePropertiesByNamespace,
+                    Arrays.asList(namespace, properties));
+        }
         int result = 0;
         Connection conn = null;
         PreparedStatement pstmt = null;
         StringBuilder sb = new StringBuilder();
         sb.append("update namespace set ");
-        sb.append(String.format("properties = '%s'", properties.toJSONString()));
+        sb.append(String.format("properties = '%s'", properties));
         sb.append(String.format(" where namespace = '%s'", namespace));
         try {
             conn = DBConnector.getConn();
@@ -149,4 +182,23 @@ public class NamespaceDao {
             DBConnector.closeConn(pstmt, conn);
         }
     }
+
+    public static Namespace namespaceFromResultSet(ResultSet rs) throws SQLException {
+        String comment = rs.getString("comment");
+        return Namespace.newBuilder()
+                .setNamespace(rs.getString("namespace"))
+                .setProperties(rs.getString("properties"))
+                .setComment(comment == null ? "" : comment)
+                .setDomain(rs.getString("domain"))
+                .build();
+    }
+
+    public static Namespace DEFAULT_NAMESPACE =
+            Namespace.newBuilder()
+                    .setNamespace(DBConfig.LAKESOUL_DEFAULT_NAMESPACE)
+                    .setProperties("{}")
+                    .setComment("")
+                    .setDomain(AuthZContext.getInstance().getDomain())
+                    .build();
+
 }
